Create a new pull request by comparing

+++ src/driver/sqlserver/SqlServerQueryRunner.ts
@@ -595,7 +595,6 @@ export class SqlServerQueryRunner extends BaseQueryRunner implements QueryRunner
             oldTableName = splittedName[1];
         }
 
-        newTable.path = this.driver.buildTableName(newTableName, newTable.schema, newTable.database);
         newTable.name = this.driver.buildTableName(newTableName, schemaName, dbName);
 
         // if we have tables with database which differs from database specified in config, we must change currently used database.
@@ -670,7 +669,6 @@ export class SqlServerQueryRunner extends BaseQueryRunner implements QueryRunner
         await this.executeQueries(upQueries, downQueries);
 
         // rename old table and replace it in cached tabled;
-        oldTable.path = newTable.path;
         oldTable.name = newTable.name;
         this.replaceCachedTable(oldTable, newTable);
     }
@@ -1416,23 +1414,64 @@ export class SqlServerQueryRunner extends BaseQueryRunner implements QueryRunner
                 ? `SELECT * FROM "${database}"."INFORMATION_SCHEMA"."TABLES" WHERE "TABLE_TYPE" = 'BASE TABLE'`
                 : `SELECT * FROM "INFORMATION_SCHEMA"."TABLES" WHERE "TABLE_TYPE" = 'BASE TABLE'`;
             const allTablesResults: ObjectLiteral[] = await this.query(allTablesSql);
-            await Promise.all(allTablesResults.map(async tablesResult => {
-                // const tableName = database ? `"${tablesResult["TABLE_CATALOG"]}"."sys"."foreign_keys"` : `"sys"."foreign_keys"`;
-                const dropForeignKeySql = `SELECT 'ALTER TABLE "${tablesResult["TABLE_CATALOG"]}"."' + OBJECT_SCHEMA_NAME("fk"."parent_object_id", DB_ID('${tablesResult["TABLE_CATALOG"]}')) + '"."' + OBJECT_NAME("fk"."parent_object_id", DB_ID('${tablesResult["TABLE_CATALOG"]}')) + '" ` +
-                    `DROP CONSTRAINT "' + "fk"."name" + '"' as "query" FROM "${tablesResult["TABLE_CATALOG"]}"."sys"."foreign_keys" AS "fk" ` +
-                    `WHERE "fk"."referenced_object_id" = OBJECT_ID('"${tablesResult["TABLE_CATALOG"]}"."${tablesResult["TABLE_SCHEMA"]}"."${tablesResult["TABLE_NAME"]}"')`;
-                const dropFkQueries: ObjectLiteral[] = await this.query(dropForeignKeySql);
-                return Promise.all(dropFkQueries.map(result => result["query"]).map(dropQuery => this.query(dropQuery)));
-            }));
-            await Promise.all(allTablesResults.map(tablesResult => {
-                if (tablesResult["TABLE_NAME"].startsWith("#")) {
-                    // don't try to drop temporary tables
-                    return;
-                }
 
-                const dropTableSql = `DROP TABLE "${tablesResult["TABLE_CATALOG"]}"."${tablesResult["TABLE_SCHEMA"]}"."${tablesResult["TABLE_NAME"]}"`;
-                return this.query(dropTableSql);
-            }));
+            if (allTablesResults.length > 0) {
+                const tablesByCatalog: { [key: string]: { TABLE_NAME: string, TABLE_SCHEMA: string }[] } = allTablesResults.reduce(
+                    (c, { TABLE_CATALOG, TABLE_SCHEMA, TABLE_NAME }) => {
+                        c[TABLE_CATALOG] = c[TABLE_CATALOG] || [];
+                        c[TABLE_CATALOG].push({ TABLE_SCHEMA, TABLE_NAME });
+                        return c;
+                    },
+                    {}
+                )
+
+                const foreignKeysSql = Object.entries(tablesByCatalog).map(([ TABLE_CATALOG, tables ]) => {
+                    const conditions = tables.map(({ TABLE_SCHEMA, TABLE_NAME }) => {
+                        return `("fk"."referenced_object_id" = OBJECT_ID('"${TABLE_CATALOG}"."${TABLE_SCHEMA}"."${TABLE_NAME}"'))`
+                    }).join(" OR ")
+
+                    return `
+                        SELECT DISTINCT '${TABLE_CATALOG}' AS                                              "TABLE_CATALOG",
+                                        OBJECT_SCHEMA_NAME("fk"."parent_object_id",
+                                                           DB_ID('${TABLE_CATALOG}')) AS                   "TABLE_SCHEMA",
+                                        OBJECT_NAME("fk"."parent_object_id", DB_ID('${TABLE_CATALOG}')) AS "TABLE_NAME",
+                                        "fk"."name" AS                                                     "CONSTRAINT_NAME"
+                        FROM "${TABLE_CATALOG}"."sys"."foreign_keys" AS "fk"
+                        WHERE (${conditions})
+                    `;
+                }).join(" UNION ALL ");
+
+                const foreignKeys: { TABLE_CATALOG: string, TABLE_SCHEMA: string, TABLE_NAME: string, CONSTRAINT_NAME: string }[] = await this.query(
+                    foreignKeysSql);
+
+                await Promise.all(foreignKeys.map(async ({
+                    TABLE_CATALOG,
+                    TABLE_SCHEMA,
+                    TABLE_NAME,
+                    CONSTRAINT_NAME
+                }) => {
+                    // Disable the constraint first.
+                    await this.query(
+                        `ALTER TABLE "${TABLE_CATALOG}"."${TABLE_SCHEMA}"."${TABLE_NAME}" ` +
+                        `NOCHECK CONSTRAINT "${CONSTRAINT_NAME}"`
+                    );
+
+                    await this.query(
+                        `ALTER TABLE "${TABLE_CATALOG}"."${TABLE_SCHEMA}"."${TABLE_NAME}" ` +
+                        `DROP CONSTRAINT "${CONSTRAINT_NAME}" -- FROM CLEAR`
+                    );
+                }));
+
+                await Promise.all(allTablesResults.map(tablesResult => {
+                    if (tablesResult["TABLE_NAME"].startsWith("#")) {
+                        // don't try to drop temporary tables
+                        return;
+                    }
+
+                    const dropTableSql = `DROP TABLE "${tablesResult["TABLE_CATALOG"]}"."${tablesResult["TABLE_SCHEMA"]}"."${tablesResult["TABLE_NAME"]}"`;
+                    return this.query(dropTableSql);
+                }));
+            }
 
             await this.commitTransaction();
 
@@ -1525,57 +1564,54 @@ export class SqlServerQueryRunner extends BaseQueryRunner implements QueryRunner
 
         if (!tableNames) {
             const databasesSql = `
-                SELECT "name" FROM "master"."dbo"."sysdatabases"
+                SELECT DISTINCT
+                    "name"
+                FROM "master"."dbo"."sysdatabases"
                 WHERE "name" NOT IN ('master', 'model', 'msdb')
             `;
             const dbDatabases: { name: string }[] = await this.query(databasesSql);
 
             const tablesSql = dbDatabases.map(({ name }) => {
                 return `
-                    SELECT
+                    SELECT DISTINCT
                         "TABLE_CATALOG", "TABLE_SCHEMA", "TABLE_NAME"
                     FROM "${name}"."INFORMATION_SCHEMA"."TABLES"
                     WHERE
                       "TABLE_TYPE" = 'BASE TABLE'
                       AND
+                      "TABLE_CATALOG" = '${name}'
+                      AND
                       ISNULL(Objectproperty(Object_id("TABLE_CATALOG" + '.' + "TABLE_SCHEMA" + '.' + "TABLE_NAME"), 'IsMSShipped'), 0) = 0
                 `;
             }).join(" UNION ALL ");
+
             dbTables.push(...await this.query(tablesSql));
         } else {
-            const dbNames = tableNames
-                .filter(tablePath => tablePath.split(".").length === 3)
-                .map(tablePath => tablePath.split(".")[0]);
-            if (this.driver.database && !dbNames.find(dbName => dbName === this.driver.database))
-                dbNames.push(this.driver.database);
+            const tableNamesByCatalog = tableNames
+                .map(tableName => this.parseTableName(tableName, currentSchema))
+                .reduce((c, { database, ...other}) => {
+                    database = database || currentDatabase;
+                    c[database] = c[database] || []
+                    c[database].push(other);
+                    return c;
+                }, {} as { [key: string]: { schema: string, name: string }[] })
 
-            const tablesCondition = tableNames.map(tableName => {
-                let [database, schema, name] = tableName.split(".");
-                // if name is empty, it means that tableName have only schema name and table name or only table name
-                if (!name) {
-                    // if schema is empty, it means tableName have only name of a table. Otherwise it means that we have "schemaName"."tableName" string.
-                    if (!schema) {
-                        name = database;
-                        schema = this.driver.options.schema || currentSchema;
+            const tablesSql = Object.entries(tableNamesByCatalog).map(([ database, tables ]) => {
+                const tablesCondition = tables
+                    .map(({ schema, name }) => {
+                        return `("TABLE_SCHEMA" = '${schema}' AND "TABLE_NAME" = '${name}')`;
+                    })
+                    .join(" OR ");
 
-                    } else {
-                        name = schema;
-                        schema = database;
-                    }
-                } else if (schema === "") {
-                    schema = this.driver.options.schema || currentSchema;
-                }
-
-
-                return `("TABLE_SCHEMA" = '${schema}' AND "TABLE_NAME" = '${name}')`;
-            }).join(" OR ");
-
-            const tablesSql = dbNames.map(dbName => {
                 return `
-                    SELECT
+                    SELECT DISTINCT
                         "TABLE_CATALOG", "TABLE_SCHEMA", "TABLE_NAME"
-                    FROM "${dbName}"."INFORMATION_SCHEMA"."TABLES"
-                    WHERE "TABLE_TYPE" = 'BASE TABLE' AND ${tablesCondition}`;
+                    FROM "${database}"."INFORMATION_SCHEMA"."TABLES"
+                    WHERE
+                          "TABLE_TYPE" = 'BASE TABLE' AND
+                          "TABLE_CATALOG" = '${database}' AND
+                          ${tablesCondition}
+                `;
             }).join(" UNION ALL ");
 
             dbTables.push(...await this.query(tablesSql));
@@ -1695,7 +1731,6 @@ export class SqlServerQueryRunner extends BaseQueryRunner implements QueryRunner
             const schema = getSchemaFromKey(dbTable, "TABLE_SCHEMA");
             table.database = dbTable["TABLE_CATALOG"];
             table.schema = dbTable["TABLE_SCHEMA"];
-            table.path = this.driver.buildTableName(dbTable["TABLE_NAME"], dbTable["TABLE_SCHEMA"], dbTable["TABLE_CATALOG"]);
             table.name = this.driver.buildTableName(dbTable["TABLE_NAME"], schema, db);
 
             const defaultCollation = dbCollations.find(dbCollation => dbCollation["NAME"] === dbTable["TABLE_CATALOG"])!;
@@ -1856,6 +1891,8 @@ export class SqlServerQueryRunner extends BaseQueryRunner implements QueryRunner
                 return new TableForeignKey({
                     name: dbForeignKey["FK_NAME"],
                     columnNames: foreignKeys.map(dbFk => dbFk["COLUMN_NAME"]),
+                    referencedDatabase: dbForeignKey["TABLE_CATALOG"],
+                    referencedSchema: dbForeignKey["REF_SCHEMA"],
                     referencedTableName: referencedTableName,
                     referencedColumnNames: foreignKeys.map(dbFk => dbFk["REF_COLUMN"]),
                     onDelete: dbForeignKey["ON_DELETE"].replace("_", " "), // SqlServer returns NO_ACTION, instead of NO ACTION
